PM: Coverity Fixes
[Defect / Enhancement / New Feature]
MobC00303614

[Problem]
Coverity errors in few PM files

[Solution]
Fix them.

[Reviewers]
Shameer

[Testing]
    [Test Cases]

    [Label/CL# Synced To]

    [Platform]

[BROADCOM INTERNAL NOTES]

Reviewed-on: http://mps-gerrit.broadcom.com/18613
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>

+++ drivers/power/bcmpmu-fg.c
@@ -2041,7 +2041,7 @@ static void bcmpmu_fg_cal_low_batt_algo(struct bcmpmu_fg_data *fg)
 	 */
 	if (fg->flags.fully_charged) {
 		cap_full_charge = (fg->capacity_info.max_design *
-				(100 + cap_delta));
+				(100 + (u64)cap_delta));
 		cap_full_charge = div_s64(cap_full_charge, 100);
 		fg->capacity_info.full_charge = cap_full_charge;
 		bcmpmu_fg_save_full_charge_cap(fg, (100 + cap_delta));
